Move C++ libs into folder algorithms
